udpate amout type : double to decimal

+++ Sources/WalletManagerDAL/Serializer/BanquePostaleCsvSerializer.cs
@@ -24,7 +24,7 @@ namespace WalletManagerDAL.Serializer
                         OperationDate = Convert.ToDateTime(values[0]),
                         Label = values[1],
                         Reference = Guid.NewGuid().ToString(),
-                        Amount = Convert.ToDouble(values[2].Replace('.', ',')),
+                        Amount = Convert.ToDecimal(values[2].Replace('.', ',')),
                         Category = WalletManagerDTO.Enumerations.TransactionCategory.NA
                     };
