+++ Sources/WalletManagerTestProject/Utils/FakerUtils.cs
@@ -6,7 +6,7 @@ namespace WalletManagerTestProject.Utils
     public static class FakerUtils
     {
         public static Faker<WalletManagerDTO.Transaction> GetTransactionDtoFaker => new Faker<WalletManagerDTO.Transaction>()
-            .RuleFor(field => field.Amount, faker => faker.Random.Double())
+            .RuleFor(field => field.Amount, faker => faker.Random.Decimal())
             .RuleFor(field => field.Category, faker => faker.Random.Enum<WalletManagerDTO.Enumerations.TransactionCategory>())
             .RuleFor(field => field.Compte, faker => faker.Random.String2(2))
             .RuleFor(field => field.Label, faker => faker.Random.String2(2))
@@ -14,7 +14,7 @@ namespace WalletManagerTestProject.Utils
             .RuleFor(field => field.Reference, faker => faker.Random.String2(2));
 
         public static Faker<WalletManagerSite.Models.TransactionViewModel> GetTransactionViewModelFaker => new Faker<WalletManagerSite.Models.TransactionViewModel>()
-            .RuleFor(field => field.Amount, faker => faker.Random.Double())
+            .RuleFor(field => field.Amount, faker => faker.Random.Decimal())
             .RuleFor(field => field.Category, faker => faker.Random.Enum<WalletManagerDTO.Enumerations.TransactionCategory>())
             .RuleFor(field => field.Compte, faker => faker.Random.String2(2))
             .RuleFor(field => field.Label, faker => faker.Random.String2(2))
